Simplify RelMemberType
This makes the code a bit more performant by removing the wrapping RelMemberType as a separate enum. No point now with protobuf3 as it gets generated as is.

I had to modify the .proto file to make the casing consistent, but eventually I do hope the protobuf lib itself can do that on the fly.