Remove wildcarded pub use statements in lib.rs
These statements put all types from the crate into the namespace
of a project that has the crate in its Cargo.toml. The re-exports
may be convenient, as you don't need to import anything from the
crate, but it is very bad practice. There are bound to be naming
conflicts, and as far as I know there is no way to prevent this
for someone who wants to use this crate.