Unify DenseNode with Node
In the OSM format, Node and DenseNode represent the same data but in
different formats. This is an OSM fileformat-specific thing (there's no
such distinction in the database or the xml format) and in older
versions this crate required the user to handle Node and DenseNode
separately. This patch unifies them, making the library easier to use.

The node handling code is split into to two conceptual parts - the "raw"
stage concerned with retrieving the raw data from a Node or DenseNode
which can be accessed via the RawNodeData Trait. To avoid creating a dyn
Trait object, we also introduce a new Enum MaybeDenseRawNode which can
either store a reference to an osmformat::Node or a DenseRawNode struct
and implements the RawNodeData trait by forwarding each method call to
the corresponding implementation on each type. This is the same
technique as used by the enum_dispatch crate, which claims it is ten
times faster than the equivalent using a dyn Trait object (but it's only
5 methods so I implemented it by hand to avoid pulling in the
  enum_dispatch crate and its dependencies). Because Nodes and
DenseNodes represent the tags in different formats (Nodes have an array
of keys and a separate array of values while DenseNodes interleave keys
and values in the same array) we need to do the same trick with
MaybeDenseRawTagIter, an enum which can either store a RawTagIter or a
DenseRawTagIter.

The second conceptual part is the crate::elements::Node struct which is
modified to store a MaybeDenseRawNode instead of an osmformat::Node and
provides all the conversions, string lookups etc for both
osmformat::Node and DenseRawNode, reducing code duplication and hiding
the details of whether a node is reprented in an osmformat::Node or
osmformat::DenseNode from the crate's user.

For the osmformat::Info (metadata) we were able to actually get rid of
the DenseNodeInfo struct entirely and make the iterator return an
osmformat::Info struct, so we don't need a MaybeDenseNodeInfo wrapper
Enum like we did for the nodes and the tags.

Aside from the user not needing to handle a separate DenseNode type with
a slightly different API, very little else has changed. A few nitpicky
details:

we change the Item type for DenseRawTagIter from (i32, i32) to
(u32, u32) to match the RawTagIter. The difference arose because
osmformat::{Node, Way, Relation} store tags as u32 indices into the
string table, but osmformat::DenseNode applies delta coding so it needs
to represent negative numbers. After decoding (which is done by the
DenseRawTagIter) the resulting indices must be positive (since a
negative index into an array is invalid anyway) so converting to a u32
is fine.

In the old implementation, Node.info() would return a reference with the
lifetime of the underlying PrimitiveBlock, while DenseNode.info would
provide a DenseNodeInfo struct which actually has the lifetime of the
DenseNode not the PrimitiveBlock (it's a bit confusing because both the
lifetime of the DenseNode and the lifetime of the PrimitiveBlock are
written as 'a in the code, but I'm about 95% sure this is correct). In
this patch, the lifetime of the Node.info() is limited by the lifetime
of the Node regardless whether it is represented as a Node or a
DenseNode. Since crate users were already dealing with this shorter
lifetime in the case of DenseNode I don't think this will cause
significant issues.